Updated product Entity services to create and deleet product detail when a product is created and deleted. Created a interface and implementation for IProductDetailDao

+++ src/main/java/com/bits/saas/dao/IProductDao.java
@@ -6,15 +6,15 @@ import com.bits.saas.exception.DaoExcpetion;
 import com.bits.saas.pojo.Product;
 
 public interface IProductDao {
-	
-int create(Product product) throws DaoExcpetion;
-	
+
+	long create(Product product) throws DaoExcpetion;
+
 	int delete(long id) throws DaoExcpetion;
-	
+
 	int update(Product product) throws DaoExcpetion;
-	
+
 	Product getProduct(long id) throws DaoExcpetion;
-	
+
 	List<Product> getProductsByEnterpriseId(long id) throws DaoExcpetion;
 
 }
